[REFACTOR]: refatoração para tornar a resposta do signatário reutilizável
