+++ src/main/java/io/github/jsousa32/libdealsign/usecases/signers/DefaultSignerCreateUseCase.java
@@ -2,8 +2,8 @@ package io.github.jsousa32.libdealsign.usecases.signers;
 
 import io.github.jsousa32.libdealsign.exceptions.DealsignException;
 import io.github.jsousa32.libdealsign.usecases.UseCase;
+import io.github.jsousa32.libdealsign.usecases.signers.models.common.SignerResponse;
 import io.github.jsousa32.libdealsign.usecases.signers.models.create.SignerCreateRequest;
-import io.github.jsousa32.libdealsign.usecases.signers.models.create.SignerCreateResponse;
 import io.github.jsousa32.libdealsign.utils.HeadersUtils;
 import io.github.jsousa32.libdealsign.utils.RestTemplateUtils;
 import org.springframework.http.HttpMethod;
@@ -11,7 +11,7 @@ import org.springframework.http.ResponseEntity;
 
 import java.util.Optional;
 
-final class DefaultSignerCreateUseCase extends UseCase<SignerCreateResponse, SignerCreateRequest> {
+final class DefaultSignerCreateUseCase extends UseCase<SignerResponse, SignerCreateRequest> {
 
     private final String bearer;
 
@@ -33,12 +33,12 @@ final class DefaultSignerCreateUseCase extends UseCase<SignerCreateResponse, Sig
     }
 
     @Override
-    public SignerCreateResponse execute(final SignerCreateRequest anInput) {
+    public SignerResponse execute(final SignerCreateRequest anInput) {
         final var rest = RestTemplateUtils.getInstance();
 
         final var httpEntity = HeadersUtils.generate(bearer, anInput);
 
-        return Optional.of(rest.exchange(this.url, HttpMethod.POST, httpEntity, SignerCreateResponse.class))
+        return Optional.of(rest.exchange(this.url, HttpMethod.POST, httpEntity, SignerResponse.class))
                 .map(ResponseEntity::getBody)
                 .orElseThrow(() -> DealsignException.generate("Não foi possível cadastrar o signatário."));
     }
