+++ src/main/java/io/github/jsousa32/libdealsign/usecases/signers/DefaultSignerService.java
@@ -1,9 +1,8 @@
 package io.github.jsousa32.libdealsign.usecases.signers;
 
+import io.github.jsousa32.libdealsign.usecases.signers.models.common.SignerResponse;
 import io.github.jsousa32.libdealsign.usecases.signers.models.create.SignerCreateRequest;
-import io.github.jsousa32.libdealsign.usecases.signers.models.create.SignerCreateResponse;
 import io.github.jsousa32.libdealsign.usecases.signers.models.update.SignerUpdateRequest;
-import io.github.jsousa32.libdealsign.usecases.signers.models.update.SignerUpdateResponse;
 
 import java.util.Objects;
 
@@ -29,12 +28,12 @@ final class DefaultSignerService implements SignerService {
     }
 
     @Override
-    public SignerCreateResponse create(final SignerCreateRequest anInput) {
+    public SignerResponse create(final SignerCreateRequest anInput) {
         return DefaultSignerCreateUseCase.generate(this.bearer, this.url).execute(anInput);
     }
 
     @Override
-    public SignerUpdateResponse update(SignerUpdateRequest anInput) {
+    public SignerResponse update(SignerUpdateRequest anInput) {
         return DefaultSignerUpdateUseCase.generate(this.bearer, this.url).execute(anInput);
     }
 }
