+++ src/main/java/io/github/jsousa32/libdealsign/usecases/signers/DefaultSignerUpdateUseCase.java
@@ -2,8 +2,8 @@ package io.github.jsousa32.libdealsign.usecases.signers;
 
 import io.github.jsousa32.libdealsign.exceptions.DealsignException;
 import io.github.jsousa32.libdealsign.usecases.UseCase;
+import io.github.jsousa32.libdealsign.usecases.signers.models.common.SignerResponse;
 import io.github.jsousa32.libdealsign.usecases.signers.models.update.SignerUpdateRequest;
-import io.github.jsousa32.libdealsign.usecases.signers.models.update.SignerUpdateResponse;
 import io.github.jsousa32.libdealsign.utils.HeadersUtils;
 import io.github.jsousa32.libdealsign.utils.RestTemplateUtils;
 import org.springframework.http.HttpMethod;
@@ -11,7 +11,7 @@ import org.springframework.http.ResponseEntity;
 
 import java.util.Optional;
 
-final class DefaultSignerUpdateUseCase extends UseCase<SignerUpdateResponse, SignerUpdateRequest> {
+final class DefaultSignerUpdateUseCase extends UseCase<SignerResponse, SignerUpdateRequest> {
 
     private final String bearer;
 
@@ -33,12 +33,12 @@ final class DefaultSignerUpdateUseCase extends UseCase<SignerUpdateResponse, Sig
     }
 
     @Override
-    public SignerUpdateResponse execute(final SignerUpdateRequest anInput) {
+    public SignerResponse execute(final SignerUpdateRequest anInput) {
         final var rest = RestTemplateUtils.getInstance();
 
         final var httpEntity = HeadersUtils.generate(bearer, anInput);
 
-        return Optional.of(rest.exchange(this.url, HttpMethod.POST, httpEntity, SignerUpdateResponse.class))
+        return Optional.of(rest.exchange(this.url, HttpMethod.POST, httpEntity, SignerResponse.class))
                 .map(ResponseEntity::getBody)
                 .orElseThrow(() -> DealsignException.generate("Não foi possível atualizar o signatário."));
     }
