--- src/main/java/io/github/jsousa32/libdealsign/usecases/signers/models/create/SignerCreateResponse.java
@@ -1,36 +1,0 @@
-package io.github.jsousa32.libdealsign.usecases.signers.models.create;
-
-import io.github.jsousa32.libdealsign.usecases.profiles.models.common.Profile;
-
-import java.time.LocalDateTime;
-import java.util.List;
-
-public class SignerCreateResponse {
-
-    private String name;
-
-    private LocalDateTime birthday;
-
-    private String document;
-
-    private List<Profile> profiles;
-
-    public SignerCreateResponse() {
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public LocalDateTime getBirthday() {
-        return birthday;
-    }
-
-    public String getDocument() {
-        return document;
-    }
-
-    public List<Profile> getProfiles() {
-        return profiles;
-    }
-}
